:construction: Rework parsers and writers into their own translation units
