+++ include/aigverse/io/write_aiger.hpp
@@ -5,13 +5,9 @@
 #ifndef AIGVERSE_WRITE_AIGER_HPP
 #define AIGVERSE_WRITE_AIGER_HPP
 
-#include <fmt/format.h>
-#include <mockturtle/io/write_aiger.hpp>
-#include <mockturtle/networks/aig.hpp>
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
+#include "aigverse/types.hpp"
 
-#include <string>
+#include <pybind11/pybind11.h>
 
 namespace aigverse
 {
@@ -20,21 +16,13 @@ namespace detail
 {
 
 template <typename Ntk>
-void write_aiger(pybind11::module& m)
-{
-    using namespace pybind11::literals;
+void write_aiger(pybind11::module& m);
 
-    m.def(
-        "write_aiger", [](const Ntk& ntk, const std::string& filename) { mockturtle::write_aiger(ntk, filename); },
-        "network"_a, "filename"_a);
-}
+extern template void write_aiger<aigverse::aig>(pybind11::module& m);
 
 }  // namespace detail
 
-inline void write_aiger(pybind11::module& m)
-{
-    detail::write_aiger<mockturtle::aig_network>(m);
-}
+void write_aiger(pybind11::module& m);
 
 }  // namespace aigverse
