+++ include/aigverse/io/write_dot.hpp
@@ -5,12 +5,9 @@
 #ifndef AIGVERSE_WRITE_DOT_HPP
 #define AIGVERSE_WRITE_DOT_HPP
 
-#include <mockturtle/io/write_dot.hpp>
-#include <mockturtle/networks/aig.hpp>
-#include <pybind11/pybind11.h>
-#include <pybind11/stl.h>
+#include "aigverse/types.hpp"
 
-#include <string>
+#include <pybind11/pybind11.h>
 
 namespace aigverse
 {
@@ -19,21 +16,13 @@ namespace detail
 {
 
 template <typename Ntk>
-void write_dot(pybind11::module& m)
-{
-    using namespace pybind11::literals;
+void write_dot(pybind11::module& m);
 
-    m.def(
-        "write_dot", [](const Ntk& ntk, const std::string& filename) { mockturtle::write_dot(ntk, filename); },
-        "network"_a, "filename"_a);
-}
+extern template void write_dot<aigverse::aig>(pybind11::module& m);
 
 }  // namespace detail
 
-inline void write_dot(pybind11::module& m)
-{
-    detail::write_dot<mockturtle::aig_network>(m);
-}
+void write_dot(pybind11::module& m);
 
 }  // namespace aigverse
